Remove shared test code for removed commands
Merge shared test class `CommandTestBase` with its only remaining using class.
